cloud_functions/forecast_assignments_past/pyproject.toml
[tool.poetry]
name = "forecast"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = ">=3.9,<3.11"
requests-cache = "^0.9.4"
pandas = "^1.4.2"
click = "^8.1.2"

pyforecast = { git = "https://github.com/urquha/py_forecast.git", branch = "master" }
google-cloud-bigquery = "^3.1.0"
google-cloud-logging = "^3.1.2"

[tool.poetry.dev-dependencies]
pytest = "^7.1.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
